+++ Magic/CHANGELOG.md
@@ -55,6 +55,7 @@
  - Add in-game notifications to ops about magic errors and warnings (can be controlled via configs or permissions)
  - Triggers can now be a simple list or a one-liner
  - Jedi and Sith will auto-learn recipes, nether stars removed (still needed for hilt)
+ - Fix auto_close on shops not working
 
 # 8.1.2